chore: move payload generation to lib/vp.js
