+++ lib/vp.js
@@ -1,6 +1,7 @@
 import { DOMAIN, VP_API_CLIENT_ID, VP_API_CLIENT_SECRET } from '../config';
 import fetch from 'node-fetch';
 import { encode } from "base-64";
+import fs from 'fs';
 
 class VP {
   constructor() {
@@ -73,6 +74,78 @@ class VP {
     return response;
   }
 
+  /**
+   * @param {DecisionmakingFlow} decisionmakingFlow
+   * @param {Pieces[]} pieces
+   */
+  generatePayload(decisionmakingFlow, pieces) {
+    const payload = {
+      '@context': [
+        'https://data.vlaanderen.be/doc/applicatieprofiel/besluitvorming/erkendestandaard/2021-02-04/context/besluitvorming-ap.jsonld',
+        {
+          'Stuk.isVoorgesteldDoor': 'https://data.vlaanderen.be/ns/dossier#isVoorgesteldDoor',
+          'Concept': 'http://www.w3.org/2004/02/skos/core#Concept',
+          'format': 'http://purl.org/dc/terms/format',
+          'content': 'http://www.w3.org/ns/prov#value',
+          'prefLabel': 'http://www.w3.org/2004/02/skos/core#prefLabel',
+          'filename': 'http://www.semanticdesktop.org/ontologies/2007/03/22/nfo#fileName',
+        }
+      ],
+      '@id': decisionmakingFlow.uri,
+      '@type': 'Besluitvormingsaangelegenheid',
+      'Besluitvormingsaangelegenheid.naam': decisionmakingFlow.name,
+      'Besluitvormingsaangelegenheid.alternatieveNaam': decisionmakingFlow.altName,
+      'Besluitvormingsaangelegenheid.beleidsveld': decisionmakingFlow.governmentFields.map(
+        (field) => ({
+          '@id': field.uri,
+          '@type': 'Concept',
+          prefLabel: field.label,
+        })
+      ),
+      '@reverse': {
+        'Dossier.isNeerslagVan': {
+          '@id': decisionmakingFlow.case,
+          '@type': 'Dossier',
+          'Dossier.bestaatUit': pieces.map(
+            (piece) => ({
+              '@id': piece.uri,
+              '@type': 'Stuk',
+              'Stuk.naam': piece.name,
+              'Stuk.creatiedatum': piece.created.toISOString(),
+              'Stuk.type': {
+                '@id': piece.type.uri,
+                '@type': 'Concept',
+                prefLabel: piece.type.label,
+              },
+              'Stuk.isVoorgesteldDoor': piece
+                .files
+                .filter((file) => fs.existsSync(file.shareUri.replace('share://', '/share/')))
+                .map((file) => {
+                const content = fs.readFileSync(
+                  file.shareUri.replace('share://', '/share/'),
+                  { encoding: 'base64' }
+                );
+                let filename = piece.name;
+                if (file.isSigned) {
+                  filename += ' (ondertekend)';
+                }
+                filename += `.${file.extension}`;
+                return {
+                  '@id': file.uri,
+                  '@type': 'http://www.w3.org/ns/dcat#Distribution',
+                  format: file.format,
+                  filename: filename,
+                  signed: file.isSigned,
+                  content,
+                }
+              })
+            })
+          ),
+        }
+      }
+    };
+    return payload;
+  }
 }
 
 export default new VP();
